delete / UserWithBoardDao and move function
